veyron/lib/cmdline: Avoid repeatedly printing default help.
The previous behavior of "help ..." would repeatedly list and
print the default help for each command and subcommand.  This
leads to cluttered output, especially when using -style=godoc to
generate package doc for each tool.  The new behavior fixes this.

Also added package doc for every binary using the cmdline
package, and fixed some minor typos.  Note that the "long"
description of each root command should be of the form "The foo
tool ..." so that the generated godoc is idiomatic.

Change-Id: Ided6ed61f569a6902a780bc552691885574b72a8