cmdline tools: Auto-generate doc.go file using "go generate".
Originally I had planned to write yet another cmdline program
that would take the directory containing the main package for
each cmdline tool, and generate the doc.go file by running help
for the tool.

However I realized we should use "go run" to run the tool, to
ensure we're running the most recent version of the code.  At
which point it's simpler just to run bash to run the tool and
generate the doc.go file.

Also fixed the help topic header, so that godoc recognizes it as
a section header.  Specifically it isn't allowed to include
parens, but is allowed to include a dash.

Change-Id: Ie1980263d42be8e85ee7c6709812f62d83df83d8