lib: support for setting manifest and pristine flag at build-time
This multi-part CL consists of related changes:

1) The v.io/x/devtools/internal/util package has been modified to
record the "current manifest", which identifies thatstate of v23
projects resulting from the latest invocation of "v23 update".

2) The v.io/x/lib/buildinfo and v.io/x/devtools/v23 have been modified
to record the current manifest information inside of the binaries that
are built using "v23 go build" or "v23 go install".

3) In addition to recording the current manifest, this CL also records
a boolean indicator of whether a binary has been built from a pristine
state (i.e. all projects are on a master branch and there are no
uncommitted changes or untracked files).

MultiPart: 2/3

Change-Id: Id5326d2bd7fe446b2ac42f765f50f992682deeb8