lib: Add stronger invariant checks for cmdline package.
It's nice if the cmdline package returns an error if a coding
invariant has been broken anywhere in the command tree, so that
simply running the program at all produces the error.

This CL adds extra checks to make sure command and topic names
are non-empty and unique.

Also remove the cmdline2/.api file; that was only left around to
allow the api check to work during the renaming CL.

Change-Id: Ia8f78c2fc41389d6044ae89b24c4d91fe2aaf1ab