lib: Add short description to godoc headers.
The auto-generated doc.go file for cmdline tools is formatted
specially so that godoc will interpret each command as a header,
which gets automatically linked to a table of contents.  This
change adds the short description from each command or help
topic, to make the TOC more useful.

Note that godoc has heurisitics for what it considers an
appropriate header; e.g. if the line contains certain puncutation
it won't be extracted as a header.  For these cases we fall back
and only use the command path, since it's better to have a less
descriptive header than to have no header at all.

Also added LineWriter.ForceVerbatim, which lets us force verbatim
mode.  This is used to ensure the header line we generate isn't
wrapped into more than one line.

MultiPart: 2/3

Change-Id: I66b4555fdf281604d384f65a51accec4f1fff542