lib: utility functions for manipulating sets of primitive type elements represented using maps
I have seen multiple implementations of string sets and integer sets
and methods the perform set operation and convert between sets and
slices. This CL provides a comprehensive set of reusable utility
functions for handling sets.

NOTE: Most of the code in this CL is auto-generated; focus your review
efforts on doc.go and gen.go.

Change-Id: If8def2d905264a26c75fa63a600dae498a42601b