x/lib/cmdline: Fix bugs regarding LookPath subcommands.
One bug caused duplicate help entries.  The symptoms of the bug
were that we'd see duplicate help output in both the short and
long descriptions, if either of the following conditions
occurred:

1) A command appeared as both a built-in Child as well as a
binary found via LookPath.

2) The user's PATH contained multiple directories that each
contained the same binary.

The fix is to dedup the subcommands.

Another bug was that if the LookPath subcommand exited with a
non-zero exit code, we wouldn't propagate that back to the
calling binary.

Change-Id: I843a8692bcb816b489ddfd7983ed8a496289445c
MultiPart: 1/2