x/lib/cmdline: Fix usage for subcommands, and short help.
It turns out the hack in v23-go/go.go was causing us to drop the
CMDLINE_PREFIX for all of the v23 binary-based subcommands.
Instead of hacking it there, a better solution is to fix up the
environment variables in gendoc.go, so that "v23 go generate"
doesn't pollute every generated doc with a "v23" prefix.

Another weirdness was that the short help for subcommands ordered
the "help" usage in between the built-in children, and the binary
subcommands, while the long help always ordered the "help" usage
at the end.  The new behavior is to put the "help" usage at the
end for both short and long help.

Change-Id: I5a75fe8eace605b69037b456a6807fb0096a5bd4
MultiPart: 1/2