"x/lib": IBE plaintext and ciphtertext must be []byte
The IBE library interface currently has the plaintext
and ciphertext as a pointer to an array of bytes of lengths
32 and 160 respectively. This does not work well for callers
of this library who would commonly have ciphertext bytes in
the form of a byte slice. (Casting from a byte slice to a
byte array incurs a copy.)

Therefore, this CL changes the interface types for plaintext
and ciphertext to []byte. The encrypt and decrypte operations
have runtime checks to ensure that the slices are of the right
size.

MultiPart: 2/5

Change-Id: Icdb7f29db762a78869b29ecc537a9516c861d2d3