lib/vlog: clarify usage of vmodule and vpath flag
It's currently hard to use these flags correctly without some
trial-and-error or looking at the implementation of the llog package.
Among points of confusion:
- whether the .go extension is part of the matched string
- what kind of pattern is expected (glob, regexp)
- whether a full match or partial match is required
This CL spells out the pattern type, and the omission of the .go
suffix. It also offers positive and negative examples since that's the
more concise and understandable way to illustrate the usage.

MultiPart: 1/2

Change-Id: Ia0f58bc07d4d2efcdfe4155954110b49fc4f1eaa