gosh: Add Cmd.CombinedOutput, and simplify stdout/stderr ordering
The Cmd.CombinedOutput method works like the exec.CombinedOutput
method; it returns the combined stdout/stderr, in the order they
were written.

The simplification to stdout/stderr ordering is to always make
writes to stdout and stderr writers synchronous, regardless of
whether we have the same underlying writers capturing both of
them.  This also has the nice property of ensuring that multiple
writers that capture both stdout and stderr will always see the
results in the same order.

Change-Id: Ib2b15092d3bd9d1566d98487965336c1d1c67aca