textutil: PrefixLineWriter should make single-line writes.
I realized that the whole point of buffering in PrefixLineWriter
is to ensure each underlying Write call contains a full line.
That ensures that if we have multiple PrefixLineWriters writing
out to os.Stdout, we won't be interleaving separate lines.

Note that if we didn't need this property, we could implement
PrefixLineWriter without any buffering or Flush.

So I've changed PrefixLineWriter to ensure single-line-Write.
Also changed to add '\n' to unterminated data when flushed.

Change-Id: I682fac1066bb9989cb232008a7c67959cb2f730a