gosh: Stop special-casing propagate output in BuildGoPkg
I'd originally overridden PropagateOutput to false in
BuildGoPkg because "go build" was spewing lots of useless
output, which interfered with debugging using
PropagateChildOutput=true, but today I realized the spew was
because of the -x flag, which we need not set. So, we stop
setting -x and drop the special-casing.

Change-Id: If17ee00c1263cc51bb27bf7bdea7caff084757f4