TBR: gosh: fix buffered_pipe's WriteTo implementation
The initial implementation failed to wait for all but the first write to
the pipe -- as soon as a Write happened, WriteTo would return.  Instead,
WriteTo ought to wait for the pipe to be Closed, and write the data from
all the Writes to the pipe that happen up until the Close.

Change-Id: I1e09261084217df5e1fd82c4309ad479bbb423f6