v.io/x/lib/cmdline: subtle bug fix for accessing parsed flags.
The commandline package has some complex logic for handling
duplicate flags. As part of this it takes a copy of the
flagset specified for the command and possibly manipulates
this copy. However, the copy is currently inaccessible to
the application and hence calls to Visit/Narg etc are
unpredictable depending on whether the original flag set
is used or not. A new member variable, ParsedFlags is added
which is guaranteed to be set to the FlagSet that is actually
parsed.

Change-Id: I676b01575a71d7e8f9049fd649647f9a3491c87e