TBR: gosh: revert v.io/c/20420
The change to move to return an error when the destination already
exists breaks vanadium-integration-test where integration tests are run
concurrently -- the logic in BuildGoPkg is inherently racy: it checks
whether the binary already exists, and proceeds to build it in a temp
location and then move it to the final location if not.  Several tests
could be running BuildGoPkg for the same binary simultaneously, and will
clash at the time when they try to move their binaries from their
respective temp dirs to the final location.

Change-Id: Ic7d8ab96555b6c96b55bae3bd5d4a9b1df732bd0