netstate: Do not default to filtering away addresses.
I'm not sure of the history behind this, but filtering out all but the
IPv4 loopback address seems to be the wrong thing to do. Overall,
the philosophy is to "find all paths".

In practice, this behavior was hurting our Android app demos,
where a device not on a public network was ending up with
only 127.0.0.1 in rpc.Server.Status().Endpoints, which in turn
meant that even though peers on the same network (or within Bluetooth
range) could see an advertisement for this service, they couldn't
connect to it over the single localhost advertised endpoint.

Change-Id: I3efca02dd9be78c57129f71e567ad7d554843663