x/lib/netconfig: fix bazel and cgo incompatibilities (#23)
This change was originally prompted by bazelbuild/rules_go#255. It still seems to be the case that bazel doesn't really allow cgo and go files to be mixed in the same package. This PR accepts that reality in a backwards compatible manner so that it can be undone should bazel's go_rules ever get fixed for cgo.