lib/cmdline: reflection/tag based flag definition (#25)
Many clients of the cmdline package end up many sub-commands and associated flags. These flags often become unwieldy as either a larger number of global variables or as specific structs defined purely for storing flags. This changes makes it possible to annotate any field in a struct with a tag (cmdline:"<flag-name>,<optional-literal-default>,<usage>) so that it can registered as flag with a specified command set. It's possible to provide non-literal and usage defaults as explained in the comments.