cmd/flagvar: allow defaults to include shell variables (#44)
This PR adds the ability to use environment variables as default values when specifying flags. This makes it easy to refer to $HOME/.config. The default value will be the expanded value, of, in this case, $HOME/.config, whereas the usage value will be $HOME/.config and not the expanded version.