trim dependencies by removing dbutil, gcm (#45)
This PR removes packages that are not used or used only once to trim this module's dependencies. dbutil was only used once by the vanadium/core, so it's moved to that repo.
gcm is no longer used and it pulls in a huge dependency on the google apis. The godepcop files are also removed since godepcop is no longer used.