cmdline/gendoc: add support for go's flag package (#52)
The go flag package writes a usage message to stderr rather than stdout. It also, depending on the version of the standard library, will exit with a status code of 2 or as of 1.15 0. This PR adds some annoying flags to handle this case.

Set -go-flag-pkg when generating doc.go for commands that use the standard flag package. It sets both of the following flags
-use-stderr - reads the usage output from stderr rather than stdout and ignores any non-zero exit codes from the sub-process.
-postprocess-output - processes the help output to remove absolute path names that refer to the temp directories used to build and run the command to be documented.