move CI from circleci to github actions. (#63)
The windows environment on github differs to that on circleci. The net effect seems to be that os.Environ returns either PATH or Path as the name of the env var for PATH. In order for the tests to pass reliably lookpath introduces a PathFromVars function that test for both uses.